# json_string_unescape: Use the returned RString as buffer